backend/requirements.txt
fastapi~=0.121.1
pydantic~=2.12.4
pip~=25.3
uvicorn[standard]~=0.30.6
firebase-admin~=6.5.0
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0

# LangChain packages
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-classic>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# LLM and embeddings
transformers>=4.30.0
tiktoken>=0.5.0

# Vector store
faiss-cpu>=1.7.4

# PDF processing (optional, for document loading)
pypdf>=3.0.0